test(work): keep profile restore regression bounded
