+++ framework/core/tests/python/test_assignment_orchestration.py
@@ -35,70 +35,6 @@ def _sha256(marker):
     return "sha256:" + marker * 64
 
 
-def test_kickoff_restores_work_control_after_dogfood_profile(monkeypatch, tmp_path):
-    active_profile = {"id": ""}
-    ensure_calls = []
-
-    monkeypatch.setattr(
-        ASSIGNMENT_CLI,
-        "_runtime",
-        lambda *_args, **_kwargs: (
-            SimpleNamespace(workspace_root=str(tmp_path)),
-            str(tmp_path / "runtime"),
-            {},
-        ),
-    )
-
-    def ensure_profile(_runtime_dir, _actor):
-        active_profile["id"] = "work-control"
-        ensure_calls.append("work-control")
-        return []
-
-    monkeypatch.setattr(ASSIGNMENT_CLI, "_ensure_profile", ensure_profile)
-    monkeypatch.setattr(
-        ASSIGNMENT_CLI.run_agent,
-        "bind_current_native_work",
-        lambda *_args: None,
-    )
-    monkeypatch.setattr(
-        ASSIGNMENT_CLI,
-        "_status",
-        lambda *_args: {
-            "phase": "claimed",
-            "assignment": {"assignment_id": "work-1"},
-        },
-    )
-
-    def consider_assignment(*_args, **_kwargs):
-        active_profile["id"] = "dogfood"
-        return {"consideration": {"receipt_root": _sha256("d")}}
-
-    monkeypatch.setattr(
-        ASSIGNMENT_CLI.dogfood_api,
-        "consider_assignment",
-        consider_assignment,
-    )
-
-    def profile_action(*_args, **_kwargs):
-        assert active_profile["id"] == "work-control"
-        return {"receipt": {"payload_hash": _sha256("a")}}
-
-    monkeypatch.setattr(ASSIGNMENT_CLI, "_profile_action", profile_action)
-
-    result = ASSIGNMENT_CLI._advance(
-        str(tmp_path),
-        False,
-        "initiative-1",
-        "work-1",
-        "executing",
-        "test-owner",
-        "test kickoff",
-    )
-
-    assert ensure_calls == ["work-control", "work-control"]
-    assert result["dogfood_consideration_root"] == _sha256("d")
-
-
 def test_assignment_atomic_paths_use_windows_extended_namespace():
     local = assignment_orchestration._filesystem_path(
         Path(r"C:\Users\Administrator\workspace\.kungfu\request.json"),
